requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
docopt==0.6.2
email-validator==1.3.1
fastapi==0.95.0
fonttools==4.39.0
frozenlist==1.3.3
future==0.18.3
get-docker-secret==1.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
line-bot-sdk==2.4.2
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.2
orjson==3.8.9
packaging==23.0
passlib==1.7.4
Pillow==9.4.0
pipreqs==0.4.11
pycparser==2.21
pydantic==1.10.6
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==10.4
yarg==0.1.9
yarl==1.8.2
python-jose==3.3.0